refactor: delete ScalarConverter ns, fix circular #37, simpler stack
