+++ src/examples/math.ts
@@ -15,8 +15,8 @@
  *     separate parser.
  *  6. Easily extendable by the addition of custom functions, variables, and more operators.
  */
-import type { Parjser } from "../lib";
-import { float, string } from "../lib";
+import { Parjser, string } from "../lib";
+import { float } from "../lib";
 import type { DelayedParjser } from "../lib/combinators";
 import { between, later, many, map, or, then, thenq } from "../lib/combinators";
