+++ src/lib/internal/combinators/exactly.ts
@@ -3,7 +3,7 @@ import { ResultKind } from "../result";
 import type { ImplicitParjser } from "../../index";
 import type { CombinatorInput } from "../combinated";
 import { Combinated } from "../combinated";
-import { wrapImplicit } from "../scalar-converter";
+import { wrapImplicit } from "../wrap-implicit";
 
 class Exactly<T> extends Combinated<T, T[]> {
     type = "exactly";
