+++ src/lib/internal/combinators/many1.ts
@@ -4,7 +4,7 @@ import type { ParjserBase } from "../parser";
 import { ResultKind } from "../result";
 import type { ParsingState } from "../state";
 import { Combinated } from "../combinated";
-import { wrapImplicit } from "../scalar-converter";
+import { wrapImplicit } from "../wrap-implicit";
 
 class Many1<T> extends Combinated<T, [T, ...T[]]> {
     type = "many1";
