+++ src/lib/internal/combinators/map.ts
@@ -3,7 +3,7 @@ import type { ParjsCombinator } from "../parjser";
 import type { ParjsProjection } from "../parjser";
 import type { ParjserBase } from "../parser";
 import { Combinated } from "../combinated";
-import { wrapImplicit } from "../scalar-converter";
+import { wrapImplicit } from "../wrap-implicit";
 
 class Map<T, S> extends Combinated<T, S> {
     type = "map";
