+++ src/lib/internal/combinators/then-pick.ts
@@ -1,7 +1,7 @@
 import type { ParjsCombinator, ParjsProjection } from "../parjser";
 import type { ParsingState } from "../state";
-import type { ImplicitParjser } from "../scalar-converter";
-import { wrapImplicit } from "../scalar-converter";
+import type { ImplicitParjser } from "../wrap-implicit";
+import { wrapImplicit } from "../wrap-implicit";
 import type { CombinatorInput } from "../combinated";
 import { Combinated } from "../combinated";
