+++ src/lib/internal/combinators/then.ts
@@ -3,7 +3,7 @@ import type { ParsingState } from "../state";
 import type { ImplicitParjser, ParjsCombinator } from "../../index";
 
 import { composeCombinator } from "./combinator";
-import { wrapImplicit } from "../scalar-converter";
+import { wrapImplicit } from "../wrap-implicit";
 import { map } from "./map";
 import type { getParsedType } from "../utils/type";
 import type { CombinatorInput } from "../combinated";
