+++ src/lib/internal/parsers/index.ts
@@ -8,11 +8,9 @@ export type { IntOptions } from "./int";
 export { int } from "./int";
 export { newline, uniNewline } from "./newline";
 export { position } from "./position";
-export { regexp } from "./regexp";
 export { rest } from "./rest";
 export { result } from "./result";
 export { state } from "./state";
-export { string } from "./string";
 export { stringLen } from "./string-len";
 export { anyStringOf } from "./string-of";
 export {
