+++ src/lib/internal/wrap-implicit.ts
@@ -1,8 +1,7 @@
 import type { Parjser } from "./parjser";
 
-import { regexp } from "./parsers/regexp";
-import { string } from "./parsers/string";
 import type { CombinatorInput } from "./combinated";
+import { regexp, string } from "./parser";
 
 /**
  * A {@link Parjser} or a literal value convertible to a {@link Parjser}.
@@ -37,19 +36,6 @@ declare global {
  * @module parjs
  */
 export type ImplicitParjser<T> = Parjser<T> | ConvertibleScalar<T>;
-
-/**
- * A helper for working with implicit parsers.
- */
-export const ScalarConverter = {
-    /**
-     * Normalizes scalars and Parjsers into Parjsers.
-     * @param scalarOrParjser The literal or parjser.
-     */
-    convert<V>(scalarOrParjser: ImplicitParjser<V>): Parjser<V> {
-        return wrapImplicit(scalarOrParjser) as Parjser<V>;
-    }
-};
 
 export function wrapImplicit<V>(scalarOrParjser: ImplicitParjser<V>): CombinatorInput<V> {
     if (typeof scalarOrParjser === "string") {
