+++ src/test/unit/combinators/maybe.spec.ts
@@ -1,5 +1,5 @@
-import type { Parjser } from "../../../lib";
-import { fail, ResultKind, string } from "../../../lib";
+import { Parjser, string } from "../../../lib";
+import { fail, ResultKind } from "../../../lib";
 import { maybe, qthen, stringify, then } from "../../../lib/combinators";
 
 describe("maybe combinator", () => {
