+++ src/test/unit/combinators/must.spec.ts
@@ -1,5 +1,5 @@
-import { ResultKind } from "../../../lib";
-import { eof, string, stringLen } from "../../../lib";
+import { ResultKind, string } from "../../../lib";
+import { eof, stringLen } from "../../../lib";
 import { must, mustCapture, or, stringify, then } from "../../../lib/combinators";
 
 describe("must combinators", () => {
