+++ src/test/unit/trace.spec.ts
@@ -1,6 +1,6 @@
 import _ from "lodash";
-import type { ParjsFailure } from "../../lib";
-import { string, whitespace } from "../../lib";
+import { ParjsFailure, string } from "../../lib";
+import { whitespace } from "../../lib";
 import { exactly, manySepBy, then } from "../../lib/combinators";
 import { visualizeTrace } from "../../lib/internal/trace-visualizer";
